이미지 파일 경로 DB에 저장하기

+++ SpringBoard/src/main/java/com/suyeon/mapper/BoardMapper.java
@@ -4,14 +4,15 @@ import java.util.List;
 
 import org.apache.ibatis.annotations.Param;
 
+import com.suyeon.dto.BoardAttachDto;
 import com.suyeon.dto.BoardDto;
 
 public interface BoardMapper {
 	public List<BoardDto> list(String castegory);
 //	parameter값을 가져올 수 없다고 오류 ㅡ> @Param으로 주기
-	public BoardDto read(@Param("num") long num, @Param("category") String category);
-	public void del(long num);
-	public void viewCount(long num);
+	public BoardDto read(@Param("num") int num, @Param("category") String category);
+	public void del(int num);
+	public void viewCount(int num);
 	public void write(BoardDto dto);
 	public void modify(BoardDto dto);
 	
@@ -25,4 +26,7 @@ public interface BoardMapper {
 	public void updateLikeCheckCancel(@Param("num") int num, @Param("user_id") String user_id);
 	public void updateLikeCancel(int num);
 	public void deleteLike(@Param("num") int num, @Param("user_id") String user_id);
+	
+
+	
 }
